Arrangement dans un unique fichier main avec arguments. Création de tsock pour gérer le puit/source
