+++ puit.h
@@ -1,0 +1,7 @@
+#ifndef HEADER_PUIT
+#define HEADER_PUIT
+
+#include "tsock.h"
+int launchPuit(void);
+
+#endif
