+++ source.h
@@ -1,0 +1,7 @@
+#ifndef HEADER_SOURCE
+#define HEADER_SOURCE
+
+#include "tsock.h"
+int launchSource(void);
+
+#endif
